refactor: テストヘルパーの責務分離
- duckdb.ts: DuckDBライフサイクル + loadCSV に集約、getDb()を削除
- fixtures.ts: layout由来のヘルパー(getAggInput, getQuestion, weightColumn)を分離
- csv.ts: buildCSV純粋関数のみ（旧csvLoader.tsからリネーム）
- 依存方向を一方向に整理: csv←generators, fixtures←layout, duckdb←csv
