+++ tests/aggregateCrossEdge.test.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect, beforeAll, afterAll } from "vitest";
 import { aggregateCross } from "../src/lib/agg/aggregateCross";
-import { setupDuckDB, teardownDuckDB, getConn, getAggInput, weightColumn } from "./helpers/duckdb";
-import { buildCSV, loadCSV } from "./helpers/csvLoader";
+import { setupDuckDB, teardownDuckDB, getConn, loadCSV } from "./helpers/duckdb";
+import { getAggInput, weightColumn } from "./helpers/fixtures";
+import { buildCSV } from "./helpers/csv";
 import type { AggInput } from "../src/lib/agg/types";
 
 beforeAll(async () => {
