+++ tests/helpers/duckdb.ts
@@ -1,8 +1,6 @@
 import { createRequire } from "node:module";
 import { readFileSync } from "node:fs";
 import { resolve } from "node:path";
-import { parseLayout, buildQuestions, findWeightColumn } from "../../src/lib/layout";
-import type { AggInput, Question } from "../../src/lib/agg/types";
 
 const require = createRequire(import.meta.url);
 const DUCKDB_DIST = resolve(require.resolve("@duckdb/duckdb-wasm"), "..");
@@ -24,12 +22,6 @@ export function getConn(): any {
   return conn;
 }
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function getDb(): any {
-  if (!db) throw new Error("setupDuckDB() has not been called yet");
-  return db;
-}
-
 export async function setupDuckDB() {
   if (conn) return;
 
@@ -46,10 +38,7 @@ export async function setupDuckDB() {
   // testdata/test_data.csv を survey ビューとして登録
   const csvPath = resolve(import.meta.dirname!, "../../testdata/test_data.csv");
   const csvText = readFileSync(csvPath, "utf-8");
-  await db.registerFileText("survey.csv", csvText);
-  await conn.query(
-    `CREATE OR REPLACE VIEW survey AS SELECT * FROM read_csv('survey.csv')`,
-  );
+  await loadCSV(csvText);
 }
 
 export async function teardownDuckDB(): Promise<void> {
@@ -59,23 +48,9 @@ export async function teardownDuckDB(): Promise<void> {
   }
 }
 
-// ── Layout-derived test helpers ──
-
-const layoutPath = resolve(import.meta.dirname!, "../../testdata/test_layout.json");
-const layout = parseLayout(readFileSync(layoutPath, "utf-8"));
-const questions = buildQuestions(layout);
-
-export function getAggInput(code: string): AggInput {
-  const q = questions.find((q) => q.code === code);
-  if (!q) throw new Error(`Question "${code}" not found in test_layout.json`);
-  const { type, columns, codes } = q;
-  return { type, columns, codes };
+export async function loadCSV(csvText: string): Promise<void> {
+  await db.registerFileText("survey.csv", csvText);
+  await conn!.query(
+    `CREATE OR REPLACE VIEW survey AS SELECT * FROM read_csv('survey.csv')`,
+  );
 }
-
-export function getQuestion(code: string): Question {
-  const q = questions.find((q) => q.code === code);
-  if (!q) throw new Error(`Question "${code}" not found in test_layout.json`);
-  return q;
-}
-
-export const weightColumn = findWeightColumn(layout);
